Split tested requirements code
